Fix scheduler dry-run crash on SMALLINT priority decode
dry_run decoded tasks.priority (SMALLINT) as i32, so sqlx rejected
every row with an INT2/INT4 mismatch and the command never produced
output. priority is i16 everywhere else (models::task); align the
tuple type.

+++ src/cli/scheduler_cmd.rs
@@ -68,7 +68,7 @@ pub async fn status(pool: &PgPool) -> Result<(), anyhow::Error> {
 /// `ygg scheduler dry-run` — print what `tick` would do without writing.
 /// Stage 1 implementation: shows what's runnable and what would be claimed.
 pub async fn dry_run(pool: &PgPool) -> Result<(), anyhow::Error> {
-    let runnable: Vec<(String, String, i32, i32)> = sqlx::query_as(
+    let runnable: Vec<(String, String, i32, i16)> = sqlx::query_as(
         r#"SELECT t.title, r.task_prefix, t.seq, t.priority
              FROM tasks t JOIN repos r USING (repo_id)
             WHERE t.runnable = TRUE
